refactor: rename table.rs to table_name.rs
